Move update channel into launcher options
