+++ pkg/launcher/options_test.go
@@ -10,6 +10,7 @@ import (
 	"time"
 
 	"github.com/kolide/kit/stringutil"
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
@@ -267,6 +268,49 @@ func getArgsAndResponse() (map[string]string, *Options) {
 	}
 
 	return args, opts
+}
+
+func TestSanitizeUpdateChannel(t *testing.T) {
+	t.Parallel()
+	var tests = []struct {
+		name            string
+		channel         string
+		expectedChannel string
+	}{
+		{
+			name:            "default",
+			expectedChannel: Stable.String(),
+		},
+		{
+			name:            "alpha",
+			channel:         Alpha.String(),
+			expectedChannel: Alpha.String(),
+		},
+		{
+			name:            "beta",
+			channel:         Beta.String(),
+			expectedChannel: Beta.String(),
+		},
+		{
+			name:            "nightly",
+			channel:         Nightly.String(),
+			expectedChannel: Nightly.String(),
+		},
+		{
+			name:            "invalid",
+			channel:         "not-a-real-channel",
+			expectedChannel: Stable.String(),
+		},
+	}
+
+	for _, tt := range tests {
+		tt := tt
+		t.Run(tt.name, func(t *testing.T) {
+			t.Parallel()
+
+			assert.Equal(t, tt.expectedChannel, SanitizeUpdateChannel(tt.channel))
+		})
+	}
 }
 
 // windowsAddExe appends ".exe" to the input string when running on Windows
